+++ Cineaste/ViewController/MovieNight/MovieNightViewController.swift
@@ -9,6 +9,7 @@
 import UIKit
 import ReSwift
 
+// swiftlint:disable:next type_body_length
 class MovieNightViewController: UITableViewController {
     @IBOutlet private weak var permissionDeniedView: UIView!
     @IBOutlet private weak var nearbyIcon: UIImageView!
@@ -32,13 +33,17 @@ class MovieNightViewController: UITableViewController {
 
     private var canUseBluetooth: Bool = true {
         didSet {
-            bluetoothIcon.tintColor = canUseBluetooth ? .accentTextOnBlack : .accentTextOnWhite
+            bluetoothIcon.tintColor = canUseBluetooth
+                ? .cineUsageActive
+                : .cineUsageInactive
         }
     }
 
     private var canUseMicrophone: Bool = true {
         didSet {
-            microphoneIcon.tintColor = canUseMicrophone ? .accentTextOnBlack : .accentTextOnWhite
+            microphoneIcon.tintColor = canUseMicrophone
+            ? .cineUsageActive
+            : .cineUsageInactive
         }
     }
 
@@ -190,9 +195,9 @@ class MovieNightViewController: UITableViewController {
     }
 
     private func configureViews() {
-        nearbyIcon.tintColor = .accentTextOnBlack
-        permissionDeniedDescription.textColor = .accentTextOnBlack
-        usageDescription.textColor = .accentTextOnBlack
+        nearbyIcon.tintColor = .cineFooter
+        permissionDeniedDescription.textColor = .cineFooter
+        usageDescription.textColor = .cineFooter
 
         let nearbyPermissionStyle = nearbyLinkPermissionDeniedTextView.paragraphStyle
         nearbyPermissionStyle.alignment = .center
@@ -206,7 +211,7 @@ class MovieNightViewController: UITableViewController {
     }
 
     private func configureTableView() {
-        tableView.backgroundColor = UIColor.basicBackground
+        tableView.backgroundColor = UIColor.cineListBackground
 
         tableView.rowHeight = UITableView.automaticDimension
         tableView.estimatedRowHeight = 80