+++ Cineaste/ViewController/Movies/MoviesViewController.swift
@@ -49,19 +49,21 @@ class MoviesViewController: UITableViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
 
-        view.backgroundColor = UIColor.basicBackground
+        view.backgroundColor = UIColor.cineListBackground
 
-        emptyListLabel.textColor = .accentTextOnBlack
-        empyListTitle.textColor = .accentTextOnBlack
+        emptyListLabel.textColor = .cineEmptyListDescription
+        empyListTitle.textColor = .cineEmptyListDescription
         empyListTitle.text = .noContentTitle
-        emptyListAddMovieButton.backgroundColor = .primaryOrange
+        emptyListAddMovieButton.backgroundColor = .cineButton
         emptyListAddMovieButton.setTitleColor(.white, for: .normal)
         emptyListAddMovieButton.setTitle(.addMovieTitle, for: .normal)
 
         emptyViewStackView.setCustomSpacing(30, after: emptyListLabel)
 
         startMovieNightButton.accessibilityLabel = .startMovieNight
+        startMovieNightButton.accessibilityIdentifier = "StartMovieNight.Button"
         addMovieButton.accessibilityLabel = .addMovieTitle
+        addMovieButton.accessibilityIdentifier = "AddMovie.Button"
 
         registerForPreviewing(with: self, sourceView: tableView)
 
@@ -121,7 +123,7 @@ class MoviesViewController: UITableViewController {
 
     private func configureTableView() {
         tableView.tableFooterView = UIView()
-        tableView.backgroundColor = UIColor.basicBackground
+        tableView.backgroundColor = UIColor.cineListBackground
 
         tableView.rowHeight = UITableView.automaticDimension
         tableView.estimatedRowHeight = 80