+++ Cineaste/ViewController/Movies/SeenMovieCell.swift
@@ -11,12 +11,14 @@ import UIKit
 class SeenMovieCell: UITableViewCell {
     static let identifier = "SeenMovieCell"
 
+    @IBOutlet weak var background: UIView!
     @IBOutlet weak var poster: UIImageView!
     @IBOutlet weak var title: UILabel!
     @IBOutlet weak var separatorView: UIView!
     @IBOutlet weak var watchedDateLabel: UILabel!
 
     func configure(with movie: Movie) {
+        background.backgroundColor = .cineCellBackground
         poster.loadingImage(from: movie.posterPath, in: .small)
 
         title.text = movie.title