+++ Cineaste/ViewController/Settings/SettingsViewController.swift
@@ -12,7 +12,14 @@ class SettingsViewController: UITableViewController {
     @IBOutlet private weak var footerView: UIView!
     @IBOutlet private weak var versionInfo: DescriptionLabel!
 
-    let settings = SettingItem.allCases
+    var settings: [SettingItem] {
+        if #available(iOS 13.0, *) {
+            return SettingItem.allCases
+        } else {
+            // Setting "Change Language" is not available pre iOS 13
+            return SettingItem.allCasesForPreIOS13
+        }
+    }
     var selectedSetting: SettingItem?
 
     var docController: UIDocumentInteractionController?
@@ -22,11 +29,11 @@ class SettingsViewController: UITableViewController {
 
         title = String.moreTitle
 
-        tableView.backgroundColor = UIColor.basicBackground
+        tableView.backgroundColor = UIColor.cineListBackground
         tableView.tableFooterView = footerView
 
         versionInfo.text = Constants.versionNumberInformation
-        versionInfo.textColor = .accentTextOnBlack
+        versionInfo.textColor = .cineFooter
     }
 
     override func viewWillAppear(_ animated: Bool) {