+++ sonar.sh
@@ -1,10 +1,5 @@
 #!/usr/bin/env bash
 
-bash \
-  ./xccov-to-sonarqube-generic.sh \
-  DerivedData/Logs/Test/*.xcresult/*_Test/action.xccovarchive/ \
-  > sonarqube-generic-coverage.xml
-
 sonar-scanner \
   -Dsonar.projectKey=spacepandas_cineaste-ios \
   -Dsonar.projectName=cineaste-ios \
@@ -17,4 +12,5 @@ sonar-scanner \
   -Dsonar.sources=Cineaste \
   -Dsonar.tests=CineasteTests,CineasteUITests \
   -Dsonar.test.inclusions=CineasteTests/**,CineasteUITests/** \
-  -Dsonar.coverageReportPaths=sonarqube-generic-coverage.xml
+  -Dsonar.log.level=WARN \
+  -Dsonar.swift.coverage.reportPath=report.llcov