+++ update-dependencies.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+
+bundle update
+pod update 
+bundle exec fastlane snapshot update